+++ R/netmetabin.R
@@ -75,7 +75,8 @@
 #'   design (only considered if \code{method = "Inverse"}).
 #' @param tol.multiarm.se A numeric for the tolerance for consistency
 #'   of standard errors in multi-arm studies which are consistent by
-#'   design (only considered if \code{method = "Inverse"}).
+#'   design (only considered if the argument is not \code{NULL} and
+#'   \code{method = "Inverse"}).
 #' @param details.chkmultiarm A logical indicating whether treatment
 #'   estimates and / or variances of multi-arm studies with
 #'   inconsistent results or negative multi-arm variances should be
@@ -396,7 +397,7 @@ netmetabin <- function(event1, n1, event2, n2,
                        tau.preset = NULL,
                        ##
                        tol.multiarm = 0.001,
-                       tol.multiarm.se = tol.multiarm,
+                       tol.multiarm.se = NULL,
                        details.chkmultiarm = FALSE,
                        details.chkdata = TRUE,
                        ##
@@ -471,7 +472,8 @@ netmetabin <- function(event1, n1, event2, n2,
     chknumeric(tau.preset, min = 0, length = 1)
   ##
   chknumeric(tol.multiarm, min = 0, length = 1)
-  chknumeric(tol.multiarm.se, min = 0, length = 1)
+  if (!is.null(tol.multiarm.se))
+    chknumeric(tol.multiarm.se, min = 0, length = 1)
   chklogical(details.chkmultiarm)
   chklogical(details.chkdata)
   ##